+++ CHANGELOG.rst
@@ -5,6 +5,6 @@ Changelog
 Version 0.x
 ===========
 
-- First release of xgb_dist
+- First release of xgboost-distribution
 - Contains XGBDistribution estimator, a xgboost wrapper with natural gradients
 - Normal distribution implemented